Add ability to set HTTP options (e.g. timeouts, etc) via config for Ethereumex.
This patch passes through an HTTP options config from the caller. This allows us to set the configuration (such as connect and recv timeout) from the caller into EthereumEx. As there are a large number of possible configuration options, it seems easier to pass the options from HTTPoison to the caller than to expose them through our own interface.